+++ parsons/solidarity_tech/exceptions.py
