+++ parsons/solidarity_tech/st_emails.py
@@ -2,8 +2,8 @@ import logging
 
 from requests.exceptions import HTTPError
 
+from parsons.solidarity_tech.exceptions import STUnexpectedResponseCodeError
 from parsons.solidarity_tech.solidarity_tech_base import SolidarityTechBase
-from parsons.solidarity_tech.solidarity_tech_exceptions import STUnexpectedResponseCodeError
 
 logger = logging.getLogger(__name__)
