+++ parsons/solidarity_tech/st_event_attendances.py
@@ -3,8 +3,8 @@ from datetime import datetime
 
 import numpy as np
 
+from parsons.solidarity_tech.exceptions import STUnexpectedResponseCodeError
 from parsons.solidarity_tech.solidarity_tech_base import SolidarityTechBase
-from parsons.solidarity_tech.solidarity_tech_exceptions import STUnexpectedResponseCodeError
 
 logger = logging.getLogger(__name__)
